[#11]Refactor: 파트너 테이블을 PartnerTable 컴포넌트로 분리
